+++ src/modules/searchBar/tests/unit/components/ActionButton.spec.js
@@ -21,8 +21,7 @@ describe("src/modules/searchBar/components/ActionButton.vue", () => {
         };
         searchBarMutations = {
             setSearchResultsActive: sinon.spy(),
-            setCurrentActionEvent: sinon.spy(),
-            setCurrentSearchInputValue: sinon.spy()
+            setCurrentActionEvent: sinon.spy()
         };
         store = createStore({
             namespaced: true,
@@ -87,7 +86,6 @@ describe("src/modules/searchBar/components/ActionButton.vue", () => {
             expect(searchBarActions.activateAction.calledOnce).to.be.true;
             expect(searchBarActions.activateAction.firstCall.args[1]).to.be.deep.equals(props);
             expect(searchBarMutations.setCurrentActionEvent.calledOnce).to.be.true;
-            expect(searchBarMutations.setCurrentSearchInputValue.calledOnce).to.be.true;
         });
 
         it("should render button with 'zoomToResult' icon", async () => {
@@ -113,7 +111,6 @@ describe("src/modules/searchBar/components/ActionButton.vue", () => {
             expect(searchBarActions.activateAction.calledOnce).to.be.true;
             expect(searchBarActions.activateAction.firstCall.args[1]).to.be.deep.equals(props);
             expect(searchBarMutations.setCurrentActionEvent.calledOnce).to.be.true;
-            expect(searchBarMutations.setCurrentSearchInputValue.calledOnce).to.be.true;
         });
 
         it("should render button with 'openGetFeatureInfo' icon", async () => {
@@ -144,7 +141,6 @@ describe("src/modules/searchBar/components/ActionButton.vue", () => {
             expect(searchBarActions.activateAction.calledOnce).to.be.true;
             expect(searchBarActions.activateAction.firstCall.args[1]).to.be.deep.equals(props);
             expect(searchBarMutations.setCurrentActionEvent.calledOnce).to.be.true;
-            expect(searchBarMutations.setCurrentSearchInputValue.calledOnce).to.be.true;
         });
 
         it("should render button with 'highlightFeature' icon", async () => {
@@ -175,7 +171,6 @@ describe("src/modules/searchBar/components/ActionButton.vue", () => {
             expect(searchBarActions.activateAction.calledOnce).to.be.true;
             expect(searchBarActions.activateAction.firstCall.args[1]).to.be.deep.equals(props);
             expect(searchBarMutations.setCurrentActionEvent.calledOnce).to.be.true;
-            expect(searchBarMutations.setCurrentSearchInputValue.calledOnce).to.be.true;
         });
 
         it("should render button with 'activateLayerInTopicTree' icon", async () => {
@@ -204,7 +199,6 @@ describe("src/modules/searchBar/components/ActionButton.vue", () => {
             expect(searchBarActions.activateAction.calledOnce).to.be.true;
             expect(searchBarActions.activateAction.firstCall.args[1]).to.be.deep.equals(props);
             expect(searchBarMutations.setCurrentActionEvent.calledOnce).to.be.true;
-            expect(searchBarMutations.setCurrentSearchInputValue.calledOnce).to.be.true;
         });
 
         it("should render button with 'addLayerToTopicTree' icon", async () => {
@@ -233,7 +227,6 @@ describe("src/modules/searchBar/components/ActionButton.vue", () => {
             expect(searchBarActions.activateAction.calledOnce).to.be.true;
             expect(searchBarActions.activateAction.firstCall.args[1]).to.be.deep.equals(props);
             expect(searchBarMutations.setCurrentActionEvent.calledOnce).to.be.true;
-            expect(searchBarMutations.setCurrentSearchInputValue.calledOnce).to.be.true;
         });
     });
     describe("methods", () => {