+++ CHANGELOG.md
@@ -37,6 +37,7 @@
     - Fixed opacity for wfs group layers.
 - WFST: The attribute names are now fully visible and no longer overlapped.
 - BuildPortal: Fixed mastercodeVersionFolderName to never include whitespaces or colons to avoid an error being thrown, if the version you are building in includes them.
+- Searchbar: The action "Show topic in folder" was fixed.
 
 ---